+++ integrations/github/github/actions/update_repo_external_custom_properties_executor.py
@@ -64,8 +64,9 @@ class UpdateRepoExternalCustomPropertiesExecutor(AbstractGithubExecutor):
             )
 
             try:
-                await self.rest_client.make_request(
-                    f"{self.rest_client.base_url}/orgs/{org}/properties/installations/values",
+                rest_client = await self.get_rest_client(org)
+                await rest_client.make_request(
+                    f"{rest_client.base_url}/orgs/{org}/properties/installations/values",
                     method="PATCH",
                     json_data={
                         "repository_names": [str(repo)],